sensors/http-zeversolar/requirements.txt
zevercloud-api
aiohttp
